Added docker file for application itself, extended docker compose with this file and nginx, separated appsetings (docker and non-docker variants)
