+++ PetFamily.Backend/docker-compose.yaml
@@ -1,0 +1,61 @@
+﻿services:
+  web:
+    container_name: main_application
+    # build:
+    #  context: .
+    #  dockerfile: src/PetFamily.Web/Dockerfile
+    image: nesterdp/pet-family-backend:1.0
+    restart: always
+    environment:
+      - ASPNETCORE_ENVIRONMENT=Docker
+    depends_on:
+      - "postgres"
+      - "seq"
+      - "minio"
+    ports:
+      - "8080:8080"
+  postgres:
+    container_name: postgres
+    image: postgres
+    restart: always
+    environment:
+      POSTGRES_DB: pet_family
+      POSTGRES_USER: postgres
+      POSTGRES_PASSWORD: postgres
+    ports:
+      - "5432:5432"
+    volumes:
+      - postgres-data:/var/lib/postgresql/data
+  seq:
+    container_name: seq
+    image: datalust/seq:latest
+    restart: always
+    environment:
+      - ACCEPT_EULA=Y
+    ports:
+      - "5341:5341"
+      - "8081:80"
+  minio:
+    container_name : minio
+    image : minio/minio
+    restart: always
+    ports:
+      - "9000:9000"
+      - "9001:9001"
+    volumes:
+      - minio_data:/data
+    command: server --console-address ":9001" /data
+  nginx:
+    container_name : nginx
+    image: nginx:latest
+    restart: always
+    ports:
+      - "80:80"
+    volumes:
+      - ./nginx.conf:/etc/nginx/conf.d/default.conf
+    depends_on:
+      - web
+
+volumes:
+  postgres-data:
+  minio_data:
