+++ PetFamily.Backend/nginx.conf
@@ -1,0 +1,8 @@
+server {
+	listen 80;
+	
+	# проксирование
+	location /api/ {
+		proxy_pass http://web:8080;
+	}
+}
