+++ PetFamily.Backend/src/Accounts/PetFamily.Accounts.Presentation/Accounts/AccountsController.cs
@@ -1,18 +1,17 @@
-using CSharpFunctionalExtensions;
-using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using PetFamily.Accounts.Application.Commands.Login;
 using PetFamily.Accounts.Application.Commands.RefreshTokens;
 using PetFamily.Accounts.Application.Commands.Register;
 using PetFamily.Accounts.Application.Queries.GetUserById;
 using PetFamily.Accounts.Contracts.Requests;
-using PetFamily.Accounts.Presentation.Requests;
+using PetFamily.Accounts.Presentation.Accounts.Requests;
 using PetFamily.Framework;
 
-namespace PetFamily.Accounts.Presentation;
+namespace PetFamily.Accounts.Presentation.Accounts;
 
 public class AccountsController : ApplicationController
 {
+    //[Permission( "accounts.GetUserInfoById")]
     [HttpGet("{id:guid}")]
     public async Task<IActionResult> GetUserInfoById(
         [FromRoute] Guid id,
@@ -67,6 +66,5 @@ public class AccountsController : ApplicationController
         //return result.ToResponse();
         //return Ok();
         return Ok(result.Value.AccessToken);
-      
     }
 }
