+++ PetFamily.Backend/src/Accounts/PetFamily.Accounts.Presentation/Accounts/Requests/RegisterUserRequest.cs
@@ -1,6 +1,6 @@
 using PetFamily.Accounts.Application.Commands.Register;
 
-namespace PetFamily.Accounts.Presentation.Requests;
+namespace PetFamily.Accounts.Presentation.Accounts.Requests;
 
 public record RegisterUserRequest(string Email, string UserName, string Password)
 {
