+++ PetFamily.Backend/src/PetFamily.Web/ApplicationConfiguration/ControllersAdder.cs
@@ -1,4 +1,4 @@
-using PetFamily.Accounts.Presentation;
+using PetFamily.Accounts.Presentation.Accounts;
 using PetFamily.Species.Presentation.Species;
 using PetFamily.Volunteers.Presentation.Volunteers;
