--- PetFamily.Backend/src/PetFamily.Web/appsettings.Development.json
@@ -1,9 +1,0 @@
-{
-  "Logging": {
-    "LogLevel": {
-      "Default": "Information",
-      "Microsoft.AspNetCore": "Warning",
-      "Microsoft.AspNetCore.HttpLogging.HttpLoggingMiddleware": "Information"
-    }
-  }
-}
