+++ PetFamily.Backend/src/PetFamily.Web/appsettings.Docker.json
@@ -6,11 +6,11 @@
     }
   },
   "ConnectionStrings": {
-    "Database": "Server=localhost;Port=5432;Database=pet_family;User Id=postgres;Password=postgres;",
-    "Seq" : "localhost:5341"
+    "Database": "Server=postgres;Port=5432;Database=pet_family;User Id=postgres;Password=postgres;",
+    "Seq" : "seq:5341"
   },
   "Minio" : {
-    "Endpoint" : "localhost:9000",
+    "Endpoint" : "minio:9000",
     "AccessKey" : "minioadmin",
     "SecretKey" : "minioadmin",
     "WithSsl" : "false"
