+++ PetFamily.Backend/src/Shared/PetFamily.Framework/ApplicationController.cs
@@ -7,7 +7,7 @@ using PetFamily.SharedKernel.CustomErrors;
 namespace PetFamily.Framework;
 
 [ApiController]
-[Route("[controller]")]
+[Route("api/[controller]")]
 public abstract class ApplicationController : ControllerBase
 {
     public override OkObjectResult Ok(object? value)
