+++ PetFamily.Backend/src/Volunteers/PetFamily.Volunteers.Infrastructure/BackgroundServices/DeleteExpiredEntitiesBackgroundService.cs
@@ -2,7 +2,6 @@ using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
-using PetFamily.Core.Files;
 using PetFamily.Core.Options;
 using PetFamily.Volunteers.Infrastructure.Services;
